kv_validate(): Pass fields to 'func's and add some default templates
!!!BACKWARDS INCOMPATIBLE CHANGE!!!

This removes the min and max fields, and instead replaces them with a
'num' template. They're mostly equivalent, but means that a template
field needs to be added when min or max are used. The returned error is
also different, as it now returns [field, 'template', 'int'] instead of
[field, 'min', 10] or something.

The documentation of kv_validate() is pretty awful.