Add tuwf->done and tuwf->pass, change error & before hook handling
This changes the way that before hooks signal whether to continue
processing or not, and is a breaking change for code that uses before
hooks with a false return value. This change does not affect the
pre_request_handler, so only code using the git version of TUWF is
affected.

This more generic control flow handling now also permits request
handlers for overlapping URL regexes, and tuwf->pass can be used to pass
control to subsequent handlers.